# backend/docs/swagger.yaml
definitions:
  internal_api_admin.CreateUserRequest:
    properties:
      allow_nesting:
        type: boolean
      cpu_allowance:
        type: integer
      cpu_quota:
        type: integer
      disk_quota:
        type: integer
      egress:
        type: integer
      ingress:
        type: integer
      io_read:
        type: integer
      io_write:
        type: integer
      ipv4_mapping_limit:
        type: integer
      ipv4_pool_limit:
        type: integer
      ipv6_mapping_limit:
        type: integer
      ipv6_pool_limit:
        type: integer
      max_cpu_per_container:
        type: integer
      memory_quota:
        type: integer
      memory_swap:
        type: boolean
      processes_limit:
        type: integer
      reverse_proxy_limit:
        type: integer
      traffic_limit:
        type: integer
      username:
        type: string
    required:
    - username
    type: object
  internal_api_admin.UpdateUserRequest:
    properties:
      allow_nesting:
        type: boolean
      cpu_allowance:
        type: integer
      cpu_quota:
        type: integer
      disk_quota:
        type: integer
      egress:
        type: integer
      ingress:
        type: integer
      io_read:
        type: integer
      io_write:
        type: integer
      ipv4_mapping_limit:
        type: integer
      ipv4_pool_limit:
        type: integer
      ipv6_mapping_limit:
        type: integer
      ipv6_pool_limit:
        type: integer
      max_cpu_per_container:
        type: integer
      memory_quota:
        type: integer
      memory_swap:
        type: boolean
      processes_limit:
        type: integer
      reverse_proxy_limit:
        type: integer
      status:
        type: string
      traffic_limit:
        type: integer
    type: object
...
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: 获取用户列表，如果提供id参数则返回用户详情
      parameters:
      - description: 用户ID（可选，提供则返回详情）
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取用户列表或详情
      tags:
      - Admin API - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户并生成密码
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_admin.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 创建用户
      tags:
      - Admin API - 用户管理
  /api/admin/users/:id:
    delete:
      consumes:
      - application/json
      description: 删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 删除用户
      tags:
      - Admin API - 用户管理
    put:
      consumes:
      - application/json
      description: 更新用户配额和状态
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_admin.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 更新用户
      tags:
      - Admin API - 用户管理
...
  /api/system/tasks:
    get:
      consumes:
      - application/json
      description: 获取任务列表，可按容器过滤
      parameters:
      - description: 容器名称（可选）
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取任务列表
      tags:
      - System API - 任务管理
  /api/system/tasks/detail:
    get:
      consumes:
      - application/json
      description: 获取指定任务的详细信息
      parameters:
      - description: 任务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取任务详情
      tags:
      - System API - 任务管理
...
swagger: "2.0"